moved legend stuff to its own folder
