+++ src/MacroTools/Augments/HeroExperienceAugment.cs
@@ -1,4 +1,5 @@
 ﻿using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using static War3Api.Common;
 
 namespace MacroTools.Augments
