+++ src/MacroTools/LegendSystem/Capital.cs
@@ -1,9 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using MacroTools.Extensions;
+using MacroTools.FactionSystem;
 using static War3Api.Common;
 
-namespace MacroTools.FactionSystem
+namespace MacroTools.LegendSystem
 {
   /// <summary>
   /// An important structure, such as the Black Temple or the Capital Palace.
