+++ src/MacroTools/QuestSystem/UtilityStructs/ObjectiveControlLegend.cs
@@ -1,5 +1,6 @@
 using MacroTools.Extensions;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
