+++ src/MacroTools/QuestSystem/UtilityStructs/ObjectiveLegendReachRect.cs
@@ -1,4 +1,5 @@
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using MacroTools.Wrappers;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
