+++ src/WarcraftLegacies.Source/GameLogic/HeroGlowFix.cs
@@ -1,5 +1,6 @@
 using MacroTools.Extensions;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using WCSharp.Events;
 using static War3Api.Common;
