+++ src/WarcraftLegacies.Source/Setup/AugmentSetup.cs
@@ -1,5 +1,6 @@
 ﻿using MacroTools.Augments;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 
 namespace WarcraftLegacies.Source.Setup
 {
