+++ src/WarcraftLegacies.Source/Setup/Legends/LegendDragonmaw.cs
@@ -1,5 +1,6 @@
 ﻿using MacroTools;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 
 namespace WarcraftLegacies.Source.Setup.Legends
 {
