+++ src/WarcraftLegacies.Source/Setup/Legends/LegendFrostwolf.cs
@@ -1,5 +1,6 @@
 using MacroTools;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using static War3Api.Common;
 
 namespace WarcraftLegacies.Source.Setup.Legends
