+++ src/WarcraftLegacies.Source/Setup/Legends/LegendGilneas.cs
@@ -1,4 +1,5 @@
 ﻿using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 
 namespace WarcraftLegacies.Source.Setup.Legends
 {
