+++ src/WarcraftLegacies.Source/Setup/Legends/LegendKulTiras.cs
@@ -1,6 +1,7 @@
 using MacroTools;
 using MacroTools.Extensions;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using static War3Api.Common;
 
 namespace WarcraftLegacies.Source.Setup.Legends
