+++ src/WarcraftLegacies.Source/Setup/Legends/LegendQuelthalas.cs
@@ -1,5 +1,6 @@
 using MacroTools;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
