+++ src/WarcraftLegacies.Source/Setup/Legends/LegendScourge.cs
@@ -1,6 +1,7 @@
 using MacroTools;
 using MacroTools.Extensions;
 using MacroTools.FactionSystem;
+using MacroTools.LegendSystem;
 using WCSharp.Shared.Data;
 using static War3Api.Common;
 using static War3Api.Blizzard;
